requirements.txt
langchain
python-dotenv
openai
requests
tiktoken
faiss-cpu
streamlit
streamlit-on-Hover-tabs
pandas
spacy